+++ examples/spot/gift_card/gift_card_token_limit.py
@@ -1,0 +1,15 @@
+#!/usr/bin/env python
+
+import logging
+from binance.spot import Spot as Client
+from binance.lib.utils import config_logging
+from examples.utils.prepare_env import get_api_key
+
+config_logging(logging, logging.DEBUG)
+
+api_key, api_secret = get_api_key()
+
+client = Client(api_key, api_secret)
+
+response = client.gift_card_token_limit(baseToken="BUSD")
+logging.info(response)